refactor: 가게평점, 메뉴 가격 관련 타입 변경

+++ src/main/java/com/idukbaduk/itseats/menu/dto/UserMenuDto.java
@@ -9,7 +9,7 @@ public class UserMenuDto {
     private Long menuId;
     private String imageUrl;
     private String name;
-    private int price;
+    private Long price;
     private String description;
-    private double rating;
+    private Double rating;
 }
